Break out each archived year into its own component
